refact: removes details page page rules and liability
